Disable `pushy_https` because it allows HTTP fallback by default!
At addition, `pushy_https` enforced HTTPS, as the name suggests, but
since Perl doesn’t support HTTPS out of the box without extra CPAN
modules, `pushy_https` was later silently altered to fallback to HTTP
with just a printed warning without inducing a failure or providing a
configurable option to induce such a failure!

When in effect, `pushy_https` overrides the `urllist`. `urllist` in my
case enforces HTTPS, so overriding with something that silently falls
back to HTTP actually decreases my security.

The default value for `pushy_https` is `undef`, which is interpreted by
`CPAN.pm` as setting it to true, so my security was silently reduced
with a Perl update without me realising it.